Update migration script to use a supported datatype

+++ webapp/db/migrate/20110724050745_create_expenses.rb
@@ -4,7 +4,7 @@ class CreateExpenses < ActiveRecord::Migration
       t.integer :id
       t.references :user
       t.string :description
-      t.double :amount
+      t.float :amount
       t.references :payment
 
       t.timestamps
